Create Translations_pt.tsv
Translation to brazilian portuguese. I tried to follow the 2 letters rule for naming the file, but it's not ideal since pt (portuguese) is for portuguese from Portugal which is almostttttt the same, but not quite (similar enough so that we can understand everything, but it will sound weird sometimes). The correct naming would be "pt-br" (portuguese from brazil), I think.

A serious problem that I found is with grammatical gender (https://en.wikipedia.org/wiki/Grammatical_gender). English is gender neutral, but languages like portuguese, french and german are not: the nouns can be masculine or feminine and the pronouns and articles (and sometimes verbs, numerals, adjectives, etc) must form an agreement with the gender of the noun.

Simple example: in english you would say "the fork" and "the knife", using the same pronoun to both, but in portuguese it would be "o garfo" ('o' because fork/garfo is a masculine noun) and "a faca" ('a' because knife/faca is a feminine noun). Yes, all knifes are females in Brazil and they WILL cut you if you don't call them by the proper pronoun. The same can be said with "a fork" and "a knife": it would be "um garfo" and "uma faca". So I think you can see the problem already. 

Unfortunately, we cant expect to translate "the" or "a" by themselves, without knowing about what they are referring (or can refer), and think that they will work properly with languages with grammatical genders.

A possible solution is to add an extra tab (extra column) that specifies the gender of an object, char or noun so that we could add an IF in the code that would add the proper pronoun or article before the noun. But I don't know if this is the best approach.

Meanwhile, this brazilian portuguese translation is as good as it gets (for the first try before many, many corrections and contributions, that is).